refactor: remove type re-exports from agg/types.ts
agg/types.ts was re-exporting types from ../types solely to shorten
import paths within agg/. Replace with direct imports from ../types.
